authorization + route grouping

+++ app/Http/Controllers/IdeaController.php
@@ -18,23 +18,33 @@ class IdeaController extends Controller
             'content' => 'required|min:5|max:240'
         ]);
 
+        $validated['user_id'] = auth()->id();
+
         Idea::create($validated);
 
         return redirect()->route('dashboard')->with('success','Idea created successfully!');
     }
 
     public function destroy(Idea $idea){
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
         $idea->delete();     
         return redirect()->route('dashboard')->with('success','Idea deleted successfully!');
     }
 
     public function edit(Idea $idea){
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
         $editing = true;
         return view('ideas.show ', compact('idea','editing'));   
     }
 
     public function update(Idea $idea){
-
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
         $validated = request()->validate([
             'content' => 'required|min:5|max:240'
         ]);
